server/pip_requirements.txt
click==7.1.2
Faker==6.6.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
protobuf==3.15.6
PyJWT==2.0.1
python-dateutil==2.8.1
six==1.15.0
text-unidecode==1.3
Werkzeug==1.0.1
